Add ability to soft delete clusters (#474)
* Add ability to soft delete clusters

Use the detach api to implement this, useful for deleting clusters that were never properly created or deleting w/o draining services if you want to leave them in-place.

* small tweaks

* small make tweak, add notice for detach

+++ cmd/plural/cd_clusters.go
@@ -67,7 +67,10 @@ func (p *Plural) cdClusterCommands() []cli.Command {
 			Usage:     "deregisters a cluster in plural cd, and drains all services (unless --soft is specified)",
 			ArgsUsage: "CLUSTER_ID",
 			Flags: []cli.Flag{
-				cli.BoolFlag{Name: "soft", Usage: "deletes a cluster in our system but doesn't drain resources, leaving them untouched"},
+				cli.BoolFlag{
+					Name:  "soft",
+					Usage: "deletes a cluster in our system but doesn't drain resources, leaving them untouched",
+				},
 			},
 		},
 		{
@@ -218,6 +221,11 @@ func (p *Plural) handleDeleteCluster(c *cli.Context) error {
 	}
 	if existing == nil {
 		return fmt.Errorf("this cluster does not exist")
+	}
+
+	if c.Bool("soft") {
+		fmt.Println("detaching cluster from Plural CD, this will leave all workloads running.")
+		return p.ConsoleClient.DetachCluster(existing.ID)
 	}
 
 	return p.ConsoleClient.DeleteCluster(existing.ID)
